........S. [ZBX-20823] fixed inability to update hosts that were created by LLD on aarch64

+++ src/libs/zbxdbhigh/template_item.c
@@ -2458,11 +2458,11 @@ static void	save_template_lld_overrides(zbx_vector_ptr_t *overrides, zbx_hashset
 			if (HOST_INVENTORY_COUNT != override_operation->inventory_mode)
 			{
 				zbx_db_insert_add_values(&db_insert_opinventory, override_operationid,
-						(int)override_operation->inventory_mode);
+						override_operation->inventory_mode);
 
 				zbx_audit_discovery_rule_update_json_add_lld_override_opinventory((*pitem)->itemid,
 						overrideid, override_operationid,
-						(int)override_operation->inventory_mode);
+						override_operation->inventory_mode);
 			}
 
 			override_operationid++;
